Don't bother testing indirectly for int128 support around here.
The condition is pointless; on the one hand, Build will have failed if it
couldn't figure out a way to to int128, so we wouldn't even get to
compiling.  Further, it doesn't work right anyway, since it's testing for
something pretending to be gcc 4.4+, whereas clang builds supporting
int128 don't necessary make that claim.